refactor: rename todoServices to todoService
